// src/laana-v0.1.1/types/action/message.ts
export interface OutgoingMessage {
    /**
     * @generated from protobuf oneof: content
     */
    content: {
        oneofKind: "bubble";
        /**
         * @generated from protobuf field: Laana.Bubble bubble = 11;
         */
        bubble: Bubble;
    } | {
        oneofKind: "file";
        /**
         * @generated from protobuf field: Laana.NormalFile file = 12;
         */
        file: NormalFile;
    } | {
        oneofKind: "singleImage";
        /**
         * @generated from protobuf field: Laana.SingleImage singleImage = 13;
         */
        singleImage: SingleImage;
    } | {
        oneofKind: "marketFace";
        /**
         * @generated from protobuf field: Laana.MarketFace marketFace = 14;
         */
        marketFace: MarketFace;
    } | {
        oneofKind: "video";
        /**
         * @generated from protobuf field: Laana.File video = 15;
         */
        video: File;
    } | {
        oneofKind: "voice";
        /**
         * @generated from protobuf field: Laana.File voice = 16;
         */
        voice: File;
    } | {
        oneofKind: "musicCard";
        /**
         * LinkCard linkCard = 17; - unable, needs signature
         * ContactCard contactCard = 18; - unable, needs signature
         *
         * @generated from protobuf field: Laana.MusicCard musicCard = 19;
         */
        musicCard: MusicCard; // LocationCard locationCard = 20; - unable, needs signature
        // ForwardMessageRef forwardMsgRef = 21; - unable, use packed messages instead
        // XmlMessage xmlMessage = 22;
        // JsonMessage jsonMessage = 23;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message Laana.SendMessagePing
 */
export interface SendMessagePing {
    /**
     * @generated from protobuf field: Laana.Peer targetPeer = 1;
     */
    targetPeer?: Peer;
    /**
     * @generated from protobuf field: Laana.OutgoingMessage message = 2;
     */
    message?: OutgoingMessage;
}
// ...
class OutgoingMessage$Type extends MessageType<OutgoingMessage> {
// ...
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: OutgoingMessage): OutgoingMessage {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* Laana.Bubble bubble */ 11:
                    message.content = {
                        oneofKind: "bubble",
                        bubble: Bubble.internalBinaryRead(reader, reader.uint32(), options, (message.content as any).bubble)
                    };
// ...
            }
        }
        return message;
    }
// ...
}
/**
 * @generated MessageType for protobuf message Laana.OutgoingMessage
 */
export const OutgoingMessage = new OutgoingMessage$Type();
// @generated message type with reflection information, may provide speed optimized methods
class SendMessagePing$Type extends MessageType<SendMessagePing> {
    constructor() {
        super("Laana.SendMessagePing", [
            { no: 1, name: "targetPeer", kind: "message", T: () => Peer },
            { no: 2, name: "message", kind: "message", T: () => OutgoingMessage }
        ]);
    }
    create(value?: PartialMessage<SendMessagePing>): SendMessagePing {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<SendMessagePing>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SendMessagePing): SendMessagePing {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* Laana.Peer targetPeer */ 1:
                    message.targetPeer = Peer.internalBinaryRead(reader, reader.uint32(), options, message.targetPeer);
                    break;
                case /* Laana.OutgoingMessage message */ 2:
                    message.message = OutgoingMessage.internalBinaryRead(reader, reader.uint32(), options, message.message);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SendMessagePing, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* Laana.Peer targetPeer = 1; */
        if (message.targetPeer)
            Peer.internalBinaryWrite(message.targetPeer, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* Laana.OutgoingMessage message = 2; */
        if (message.message)
            OutgoingMessage.internalBinaryWrite(message.message, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message Laana.SendMessagePing
 */
export const SendMessagePing = new SendMessagePing$Type();